+++ core/transfer/src/main/java/org/eclipse/dataspaceconnector/transfer/core/transfer/TransferProcessManagerImpl.java
@@ -350,6 +350,7 @@ public class TransferProcessManagerImpl implements TransferProcessManager, Provi
             } else {
                 // Process is not finished yet, so it stays in the IN_PROGRESS state
                 monitor.info(format("Transfer process %s not COMPLETED yet. The process will not advance to the COMPLETED state.", process.getId()));
+                breakLease(process);
                 return false;
             }
         }
@@ -583,8 +584,7 @@ public class TransferProcessManagerImpl implements TransferProcessManager, Provi
     private boolean processConsumerRequest(TransferProcess process, DataRequest dataRequest) {
 
         if (sendRetryManager.shouldDelay(process)) {
-            // Break lease
-            transferProcessStore.update(process);
+            breakLease(process);
             return false;
         }
 
@@ -631,6 +631,10 @@ public class TransferProcessManagerImpl implements TransferProcessManager, Provi
     private void updateTransferProcess(TransferProcess transferProcess, Consumer<TransferProcessListener> observe) {
         observable.invokeForEach(observe);
         transferProcessStore.update(transferProcess);
+    }
+
+    private void breakLease(TransferProcess process) {
+        transferProcessStore.update(process);
     }
 
     public static class Builder {