+++ extensions/api/data-management/contractnegotiation/src/main/java/org/eclipse/dataspaceconnector/api/datamanagement/contractnegotiation/ContractNegotiationApi.java
@@ -19,7 +19,9 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import org.eclipse.dataspaceconnector.api.datamanagement.contractnegotiation.model.ContractAgreementDto;
 import org.eclipse.dataspaceconnector.api.datamanagement.contractnegotiation.model.ContractNegotiationDto;
+import org.eclipse.dataspaceconnector.api.datamanagement.contractnegotiation.model.NegotiationId;
 import org.eclipse.dataspaceconnector.api.datamanagement.contractnegotiation.model.NegotiationInitiateRequestDto;
+import org.eclipse.dataspaceconnector.api.datamanagement.contractnegotiation.model.NegotiationState;
 import org.eclipse.dataspaceconnector.spi.query.SortOrder;
 
 import java.util.List;
@@ -32,11 +34,11 @@ public interface ContractNegotiationApi {
 
     ContractNegotiationDto getNegotiation(String id);
 
-    String getNegotiationState(String id);
+    NegotiationState getNegotiationState(String id);
 
     ContractAgreementDto getAgreementForNegotiation(String negotiationId);
 
-    String initiateContractNegotiation(@Valid NegotiationInitiateRequestDto initiateDto);
+    NegotiationId initiateContractNegotiation(@Valid NegotiationInitiateRequestDto initiateDto);
 
     void cancelNegotiation(String id);