+++ extensions/azure/data-plane/storage/src/main/java/org/eclipse/dataspaceconnector/azure/dataplane/azurestorage/pipeline/AzureStorageDataSink.java
@@ -14,7 +14,7 @@
 
 package org.eclipse.dataspaceconnector.azure.dataplane.azurestorage.pipeline;
 
-import org.eclipse.dataspaceconnector.azure.dataplane.azurestorage.adapter.BlobAdapterFactory;
+import org.eclipse.dataspaceconnector.azure.blob.core.api.BlobStoreApi;
 import org.eclipse.dataspaceconnector.dataplane.spi.pipeline.DataSource;
 import org.eclipse.dataspaceconnector.dataplane.spi.pipeline.ParallelSink;
 import org.eclipse.dataspaceconnector.spi.response.StatusResult;
@@ -30,10 +30,13 @@ import static org.eclipse.dataspaceconnector.spi.response.ResponseStatus.ERROR_R
  * Writes data into an Azure storage container.
  */
 public class AzureStorageDataSink extends ParallelSink {
+    // Name of the empty blob used to indicate completion. Used by consumer-side status checker.
+    public static final String COMPLETE_BLOB_NAME = ".complete";
+
     private String accountName;
     private String containerName;
     private String sharedKey;
-    private BlobAdapterFactory blobAdapterFactory;
+    private BlobStoreApi blobStoreApi;
 
     /**
      * Writes data into an Azure storage container.
@@ -42,7 +45,7 @@ public class AzureStorageDataSink extends ParallelSink {
         for (DataSource.Part part : parts) {
             String blobName = part.name();
             try (var input = part.openStream()) {
-                try (var output = blobAdapterFactory.getBlobAdapter(accountName, containerName, blobName, sharedKey)
+                try (var output = blobStoreApi.getBlobAdapter(accountName, containerName, blobName, sharedKey)
                         .getOutputStream()) {
                     try {
                         input.transferTo(output);
@@ -57,6 +60,18 @@ public class AzureStorageDataSink extends ParallelSink {
             }
         }
         return StatusResult.success();
+    }
+
+    @Override
+    protected StatusResult<Void> complete() {
+        try {
+            // Write an empty blob to indicate completion
+            blobStoreApi.getBlobAdapter(accountName, containerName, COMPLETE_BLOB_NAME, sharedKey)
+                    .getOutputStream().close();
+        } catch (Exception e) {
+            return getTransferResult(e, "Error creating blob %s on account %s", COMPLETE_BLOB_NAME, accountName);
+        }
+        return super.complete();
     }
 
     @NotNull
@@ -90,8 +105,8 @@ public class AzureStorageDataSink extends ParallelSink {
             return this;
         }
 
-        public Builder blobAdapterFactory(BlobAdapterFactory blobAdapterFactory) {
-            sink.blobAdapterFactory = blobAdapterFactory;
+        public Builder blobStoreApi(BlobStoreApi blobStoreApi) {
+            sink.blobStoreApi = blobStoreApi;
             return this;
         }
 
@@ -99,7 +114,7 @@ public class AzureStorageDataSink extends ParallelSink {
             Objects.requireNonNull(sink.accountName, "accountName");
             Objects.requireNonNull(sink.containerName, "containerName");
             Objects.requireNonNull(sink.sharedKey, "sharedKey");
-            Objects.requireNonNull(sink.blobAdapterFactory, "blobAdapterFactory");
+            Objects.requireNonNull(sink.blobStoreApi, "blobStoreApi");
         }
 
         private Builder() {