+++ CHANGELOG.md
@@ -24,12 +24,14 @@ in the detailed section referring to by linking pull requests or issues.
 * Implement Catalog service for Data Management API (#1195)
 * Add strict body validation for REST endpoints (#1128)
 * Dependency injection using factory/provider methods (#1056)
+* Provisioned resource information in Data Management API (#1221)
 
 #### Changed
 
 * Restructure sql extension folder tree (#1154)
 * Extract single `PolicyArchive` implementation (#1158)
 * Replace `accessPolicy` and `contractPolicy` with `accessPolicyId` and `contractPolicyId` on `ContractDefinition` (#1144)
+* All DMgmt Api methods now produce and consume `APPLICATION_JSON` (#1175)
 
 #### Removed
 
@@ -42,6 +44,8 @@ in the detailed section referring to by linking pull requests or issues.
 * Fix problem with interpreting contractId/negotiationId (#1140)
 * Fixed DMgmtApi content types (#1126)
 * Fix HTTPS termination in Jetty (#1133)
+* Break lease after TransferProcessManager status check (#1214)
+* Fix path conflicts between `CatalogApiController` and `FederatedCatalogApiController` (#1225)
 
 ## [milestone-3] - 2022-04-08